+++ src/lib.rs
@@ -43,10 +43,10 @@
 
 #[macro_use]
 extern crate strum_macros;
-#[cfg(all(test, any(feature = "trait_serde_json", feature = "trait_serde_yaml")))]
+#[cfg(all(test, any(feature = "trait_serde_json", feature = "trait_serde_yaml", feature = "trait_json")))]
 #[macro_use]
 extern crate lazy_static;
-#[cfg(all(test, any(feature = "trait_serde_json", feature = "trait_serde_yaml")))]
+#[cfg(all(test, any(feature = "trait_serde_json", feature = "trait_serde_yaml", feature = "trait_json")))]
 #[cfg(test)]
 #[macro_use]
 extern crate serde_json;