+++ src/traits/mod.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "trait_json")]
+pub mod _json;
 #[cfg(feature = "trait_serde_json")]
 pub mod _serde_json;
 #[cfg(feature = "trait_serde_yaml")]