+++ CHANGELOG.md
@@ -2,5 +2,7 @@ Changelog
 =========
 
 0.0.0 (not released)
---------------------
-- Initial project skeleton
+------------------
+- Initial project release
+- Definition of `JsonType` trait
+- Implementation of trait for [`json::JsonValue`](https://github.com/maciejhirsz/json-rust/), [`serde_json::Value`](https://github.com/serde-rs/json/) and [`serde_yaml::Value`](https://github.com/dtolnay/serde-yaml).